requirements.txt
yfinance==0.2.40
gspread==6.1.2
google-auth==2.29.0
